+++ Freeform.Python/FreeformRigging/Maya/rigging/rig_overdrivers/__init__.py
@@ -1,0 +1,8 @@
+import pkgutil
+import sys
+import inspect
+
+for loader, name, is_pkg in pkgutil.walk_packages(__path__):
+	if not is_pkg:
+		module = loader.find_module(name).load_module(name)
+		setattr(sys.modules[__package__], name, module)
